Fix CLI crash on empty policy set
CLI would crash if the zero value of PolicySet was used in the pretty
printer. This is because the pretty printer expected zero values to be
printable strings, not pointers.

To fix this the embedded policy type was embedded as a struct not a
pointer to struct, avoiding the nil pointer dereference error.